Finished the Bootstrap folders restructure.
