+++ generators/app/constants.js
@@ -1,2 +1,15 @@
 module.exports.PUSH_NOTIFICATIONS_SETUP_LINK =
   'https://github.com/react-native-community/react-native-push-notification-ios';
+
+module.exports.GENERATOR_FEATURES = [
+  'Bitrise',
+  'Crashlytics',
+  'Drawer',
+  'Firebase Analytics',
+  'Firebase Performance',
+  'Login and SignUp',
+  'OnBoarding',
+  'Push Notifications',
+  'Redux Persist',
+  'Tabs'
+];
