+++ generators/app/tasks/appSetup/coreFiles/androidProjectSetup.js
