+++ generators/app/tasks/appSetup/coreFiles/appIcons.js
