+++ generators/app/tasks/appSetup/coreFiles/babelConfigSetup.js
