+++ generators/app/tasks/appSetup/coreFiles/cleanTargetsFromPods.js
