+++ generators/app/tasks/appSetup/coreFiles/iosProjectSetup.js
