+++ generators/app/tasks/appSetup/coreFiles/multipleEnvIos/schemeBase.js
