+++ generators/app/tasks/appSetup/coreFiles/packageJsonScripts.js
