+++ generators/app/tasks/appSetup/coreFiles/prettierrcConfigSetup.js
