+++ generators/app/tasks/appSetup/coreFiles/splashScreenSetup.js
