+++ generators/app/tasks/appSetup/coreFiles/tabletSetup.js
