+++ generators/app/tasks/appSetup/featuresFiles/disableLandscapeOrientation.js
