+++ generators/app/tasks/appSetup/featuresFiles/firebaseCoreFeatureFiles.js
