+++ generators/app/tasks/appSetup/featuresFiles/firebasePerformanceSetup.js
