+++ generators/app/tasks/appSetup/featuresFiles/pushNotificationsSetup.js
