+++ generators/app/tasks/configuringTasks/addFilesToGitIgnore.js
