+++ generators/app/tasks/configuringTasks/reactNativeCliInstall.js
@@ -1,4 +1,4 @@
-const runCommand = require('./runCommand');
+const runCommand = require('../runCommand');
 
 module.exports = function reactNativeCliInstall() {
   return runCommand({
