+++ generators/app/tasks/configuringTasks/reactNativeInit.js
@@ -1,4 +1,4 @@
-const runCommand = require('./runCommand');
+const runCommand = require('../runCommand');
 
 module.exports = function reactNativeInit() {
   return runCommand({
