+++ generators/app/tasks/installTasks/installPods.js
@@ -1,4 +1,4 @@
-const runCommand = require('./runCommand');
+const runCommand = require('../runCommand');
 
 module.exports = function installPods() {
   return runCommand({
