+++ generators/app/tasks/installTasks/scriptIosConfig.rb
